Update the developer guide to reflect current practice
The biggest change is to update the release procedure to reflect the PYPI automation as well as add instructions for conda packages on conda-forge. This also updates to include more services we use like AppVeyor, Code Climate, and Scrutinizer, as well as fixes a few formatting errors and typos. [ci skip]